app/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==21.1.0
packaging==23.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
